+++ bounded_contexts/reward_points/presentation/dependencies.py
@@ -19,6 +19,9 @@ from bounded_contexts.reward_points.application.use_cases.accept_invitation impo
 from bounded_contexts.reward_points.application.use_cases.add_child_membership import (
     AddChildMembershipUseCase,
 )
+from bounded_contexts.reward_points.application.use_cases.approve_independence import (
+    ApproveIndependenceUseCase,
+)
 from bounded_contexts.reward_points.application.use_cases.create_family import CreateFamilyUseCase
 from bounded_contexts.reward_points.application.use_cases.dissolve_family import DissolveFamilyUseCase
 from bounded_contexts.reward_points.application.use_cases.ensure_user_can_be_deleted import (
@@ -28,6 +31,10 @@ from bounded_contexts.reward_points.application.use_cases.issue_invitation impor
 from bounded_contexts.reward_points.application.use_cases.leave_family import LeaveFamilyUseCase
 from bounded_contexts.reward_points.application.use_cases.list_families import ListFamiliesUseCase
 from bounded_contexts.reward_points.application.use_cases.list_invitations import ListInvitationsUseCase
+from bounded_contexts.reward_points.application.use_cases.propose_independence import (
+    ProposeIndependenceUseCase,
+    RevokeIndependenceProposalUseCase,
+)
 from bounded_contexts.reward_points.application.use_cases.record_point_transaction import (
     RecordPointTransactionUseCase,
 )
@@ -173,6 +180,26 @@ def get_dissolve_family_use_case(
     return DissolveFamilyUseCase(access, families, memberships)
 
 
+def get_propose_independence_use_case(access: AccessDep, memberships: MembershipRepoDep) -> ProposeIndependenceUseCase:
+    return ProposeIndependenceUseCase(access, memberships)
+
+
+def get_revoke_independence_use_case(
+    access: AccessDep, memberships: MembershipRepoDep
+) -> RevokeIndependenceProposalUseCase:
+    return RevokeIndependenceProposalUseCase(access, memberships)
+
+
+def get_approve_independence_use_case(
+    access: AccessDep,
+    memberships: MembershipRepoDep,
+    ledgers: LedgerRepoDep,
+    transactions: TransactionRepoDep,
+    provisioning: ProvisioningDep,
+) -> ApproveIndependenceUseCase:
+    return ApproveIndependenceUseCase(access, memberships, ledgers, transactions, provisioning)
+
+
 def get_add_child_use_case(
     access: AccessDep, memberships: MembershipRepoDep, ledgers: LedgerRepoDep
 ) -> AddChildMembershipUseCase:
@@ -265,6 +292,9 @@ ViewFamilyDep = Annotated[ViewFamilyUseCase, Depends(get_view_family_use_case)]
 RenameFamilyDep = Annotated[RenameFamilyUseCase, Depends(get_rename_family_use_case)]
 LeaveFamilyDep = Annotated[LeaveFamilyUseCase, Depends(get_leave_family_use_case)]
 DissolveFamilyDep = Annotated[DissolveFamilyUseCase, Depends(get_dissolve_family_use_case)]
+ProposeIndependenceDep = Annotated[ProposeIndependenceUseCase, Depends(get_propose_independence_use_case)]
+RevokeIndependenceDep = Annotated[RevokeIndependenceProposalUseCase, Depends(get_revoke_independence_use_case)]
+ApproveIndependenceDep = Annotated[ApproveIndependenceUseCase, Depends(get_approve_independence_use_case)]
 AddChildDep = Annotated[AddChildMembershipUseCase, Depends(get_add_child_use_case)]
 RemoveMembershipDep = Annotated[RemoveMembershipUseCase, Depends(get_remove_membership_use_case)]
 IssueInvitationDep = Annotated[IssueInvitationUseCase, Depends(get_issue_invitation_use_case)]
@@ -283,18 +313,21 @@ __all__ = [
     "AcceptInvitationDep",
     "AccessDep",
     "AddChildDep",
+    "ApproveIndependenceDep",
     "CreateFamilyDep",
     "DissolveFamilyDep",
     "IssueInvitationDep",
     "LeaveFamilyDep",
     "ListFamiliesDep",
     "ListInvitationsDep",
+    "ProposeIndependenceDep",
     "RecordTransactionDep",
     "RedeemInvitationDep",
     "RemoveMembershipDep",
     "RenameFamilyDep",
     "ResetChildPasswordDep",
     "ReverseTransactionDep",
+    "RevokeIndependenceDep",
     "RevokeInvitationDep",
     "SuggestReasonsDep",
     "ViewFamilyDep",